% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_spectrum.R
\name{assign_spectrum}
\alias{assign_spectrum}
\title{assign a table of fragment masses to a spectrum data object}
\usage{
assign_spectrum(spectrum = NULL, peptide = NULL, tolerance = 0.1, ...)
}
\arguments{
\item{spectrum}{a spectrum data object}

\item{tolerance}{a numeric}

\item{table_fragments}{a fragment data object}
}
\value{
a tibble
}
\description{
\code{assign_spectrum()} converts an amino acid character to a numeric
}
