% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_string.R
\name{assign_string}
\alias{assign_string}
\title{assign a table of fragment masses to a spectrum data object}
\usage{
assign_string(spectrum = NULL, table_fragments = NULL, tolerance = 0.1)
}
\arguments{
\item{spectrum}{a spectrum data object}

\item{table_fragments}{a fragment data object}

\item{tolerance}{a numeric}
}
\value{
a list
}
\description{
\code{assign_string()} converts an amino acid character to a numeric
}
