% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobject.R
\name{check_ms2spectra}
\alias{check_ms2spectra}
\title{Check the integrity of a msfastar data object}
\usage{
check_ms2spectra(x = NULL)
}
\arguments{
\item{data}{msfastar data object}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{check_fasta()} is a helper function that checks the structure and contents of
a msfastar data object
}
