% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_precursor}
\alias{filter_precursor}
\title{Return the top n values from a vector}
\usage{
filter_precursor(df, mz)
}
\arguments{
\item{df}{dataframe}

\item{n}{n indexes}
}
\value{
vector
}
\description{
\code{filter_topn()} get the top n values from a vector
}
