% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_topn}
\alias{filter_topn}
\title{Return the top n values from a vector}
\usage{
filter_topn(df, n)
}
\arguments{
\item{df}{dataframe}

\item{n}{n indexes}
}
\value{
vector
}
\description{
\code{filter_topn()} get the top n values from a vector
}
