% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_xcorr}
\alias{filter_xcorr}
\title{Xcorr normalized spectrum}
\usage{
filter_xcorr(spectrum = NULL, chunk = 10, norm = 50)
}
\arguments{
\item{spectrum}{The spectra data object}

\item{chunk}{The number of divisions to normalize across}

\item{norm}{The value to normalize to}
}
\description{
\code{filter_xcorr()} get the top n values from a vector
}
