% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_unique.R
\name{id_unqiue}
\alias{id_unqiue}
\title{create a unique id value}
\usage{
id_unqiue(n = 1, length = 5)
}
\arguments{
\item{n}{a spectrum data object}

\item{length}{a fragment data object}
}
\value{
a character vector
}
\description{
\code{id_unqiue()} converts an amino acid character to a numeric
}
