% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{indef_fragments}
\alias{indef_fragments}
\title{Return the fragment indexes}
\usage{
indef_fragments(seq)
}
\arguments{
\item{seq}{The peptide sequence string}

\item{tolerance}{The numerical float for mass tolerance in Th}
}
\description{
Return the fragment indexes
}
\examples{
index_fragments('SA[M15.99]PLER')
}
