% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mass_atomic}
\alias{mass_atomic}
\title{Return the mass of an atom}
\usage{
mass_atomic(atom)
}
\arguments{
\item{atom}{The IUPAC amomic letter designation}
}
\description{
Return the mass of an atom
}
\examples{
mass_atomic('C')
mass_atomic('Na')
}
