% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mass_charged}
\alias{mass_charged}
\title{Return the charged mass}
\usage{
mass_charged(mass, z)
}
\arguments{
\item{mass}{The floating point mass value}

\item{z}{The integer charge value}
}
\description{
Return the charged mass
}
\examples{
mass_charged(1234.567, 2)
}
