% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mass_fragments}
\alias{mass_fragments}
\title{Return a simplified fragment mz vector}
\usage{
mass_fragments(seq)
}
\arguments{
\item{seq}{The peptide sequence string}
}
\description{
Return a simplified fragment mz vector
}
\examples{
mass_fragments('SA[M15.99]PLER')
}
