% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mass_neutral}
\alias{mass_neutral}
\title{Return the neutral mass}
\usage{
mass_neutral(mz, z)
}
\arguments{
\item{mz}{The floating point mass value}

\item{z}{The integer charge value}
}
\description{
Return the neutral mass
}
\examples{
mass_neutral(1234.567, 2)
}
