% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mass_residue}
\alias{mass_residue}
\title{Return the mass of a amino acid residue}
\usage{
mass_residue(s)
}
\arguments{
\item{s}{The single letter designation for an amino acid}
}
\description{
Return the mass of a amino acid residue
}
\examples{
mass_residue('G')
mass_residue('Y')
}
