% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_isotopes.R
\name{model_isotopes}
\alias{model_isotopes}
\title{Convert a peptide string to a named variable}
\usage{
model_isotopes(
  averagene = c("N", "A", "R", "D", "C", "E", "Q", "G", "H", "I", "K", "M", "F", "P",
    "S", "T", "W", "Y", "V"),
  max_n = 42
)
}
\arguments{
\item{averagene}{amino acid to use for a base polypeptide}
}
\value{
data list
}
\description{
\code{model_isotopes()} Generates a named variable
}
