% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_ms_amanda.R
\name{ms_amanda_peptide}
\alias{ms_amanda_peptide}
\title{helper function to read in platform specific results}
\usage{
ms_amanda_peptide(sequence = NULL, modifications = NULL)
}
\arguments{
\item{sequence}{string}

\item{modifications}{string}
}
\value{
string
}
\description{
helper function to read in platform specific results
}
