% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{mspredictr-package}
\alias{mspredictr}
\alias{mspredictr-package}
\title{mspredictr: A Lightweight Package for Predicting Peptide Fragment Mass Values and Plotting Mass Spectra with Assignments}
\description{
An R package for calculating peptide sequence and fragment masses relevant to mass spectrometry, along with some peptide string manipulation tools. Some of the functions with mass calculations and string manipulations have been written in Rust for a speed advantage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeffsocal.github.io/mspredictr/}
}

}
\author{
\strong{Maintainer}: Jeff Jones \email{jeffj@caltech.edu} (\href{https://orcid.org/0000-0002-7142-2222}{ORCID})

}
