% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_peptide.R
\name{munge_seq}
\alias{munge_seq}
\title{non-iteratable helper function to clean up comet sequence strings}
\usage{
munge_seq(sequence = NULL)
}
\arguments{
\item{sequence}{as character string}
}
\value{
a string
}
\description{
\code{munge_seq()} get the mass of a poly amino acid
}
