% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_omssa.R
\name{omssa_peptide}
\alias{omssa_peptide}
\title{helper function to read in platform specific results}
\usage{
omssa_peptide(sequence = NULL, modifications = NULL)
}
\arguments{
\item{sequence}{string}

\item{modifications}{string}
}
\value{
string
}
\description{
helper function to read in platform specific results
}
