% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_path.R
\name{path_to_example}
\alias{path_to_example}
\title{Helper function for displaying path to data}
\usage{
path_to_example()
}
\value{
print the table to console
}
\description{
Helper function for displaying path to data
}
