% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{peptide_length}
\alias{peptide_length}
\title{Return the unit length of a peptide sequence}
\usage{
peptide_length(sequences)
}
\arguments{
\item{sequences}{A vector of peptide sequence strings.}
}
\description{
Return the unit length of a peptide sequence
}
\examples{
peptide_mass(c('SA[M15.99]PLE', 'SAM[15.99]PLE'))
}
