% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{peptide_mass}
\alias{peptide_mass}
\title{Return the mass of a peptide sequence}
\usage{
peptide_mass(sequences)
}
\arguments{
\item{sequences}{A vector of peptide sequence strings.}
}
\description{
Return the mass of a peptide sequence
}
\examples{
peptide_mass(c('SAMPLE', 'SILLY'))
}
