% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{peptide_reverse}
\alias{peptide_reverse}
\title{Return the peptide sequence reversed}
\usage{
peptide_reverse(sequences)
}
\arguments{
\item{sequences}{A vector of peptide sequence strings.}
}
\description{
Return the peptide sequence reversed
}
\examples{
peptide_mass(c('THINK', 'SAM[15.99]PLER'))
}
