% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{peptide_xleucine}
\alias{peptide_xleucine}
\title{Return a I/L substituted sequence}
\usage{
peptide_xleucine(s)
}
\arguments{
\item{s}{The peptide sequence string.}
}
\description{
Return a I/L substituted sequence
}
\examples{
peptide_xleucine('SILLY')
}
