% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_spectrum.R
\name{prep_spectrum}
\alias{prep_spectrum}
\title{helper function to read in platform specific results}
\usage{
prep_spectrum(x, filter = FALSE)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
