% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ms2spectra}
\alias{print.ms2spectra}
\title{Tidy-Quant data object print definition}
\usage{
\method{print}{ms2spectra}(x, ...)
}
\arguments{
\item{x}{rmfasta data object}

\item{...}{unused legacy}
}
\value{
print object summary
}
\description{
Tidy-Quant data object print definition
}
