% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_comet.R
\name{read_comet}
\alias{read_comet}
\title{helper function to read in platform specific results}
\usage{
read_comet(x, cpus = 1)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
