% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mfg.R
\name{read_mgf}
\alias{read_mgf}
\title{The main function for parsing a fasta file}
\usage{
read_mgf(file_path = NULL)
}
\arguments{
\item{file_path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{read_mgf()} get the current regex
}
