% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_amanda.R
\name{read_ms_amanda}
\alias{read_ms_amanda}
\title{helper function to read in platform specific results}
\usage{
read_ms_amanda(x, cpus = 1)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
