% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{read_mzml}
\alias{read_mzml}
\title{The main function for parsing a fasta file}
\usage{
read_mzml(file_path = NULL)
}
\arguments{
\item{file_path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{read_mzml()} get the current regex
}
