% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omssa.R
\name{read_omssa}
\alias{read_omssa}
\title{helper function to read in platform specific results}
\usage{
read_omssa(x, cpus = 1)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
