% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psms.R
\name{read_psms}
\alias{read_psms}
\title{Table of PSMs results}
\usage{
read_psms(
  path_to_psms = NULL,
  path_to_mzml = NULL,
  platform = c("comet", "ms_amanda", "omssa", "sage", "xtandem"),
  cpus = 4
)
}
\arguments{
\item{platform}{as character string}

\item{path}{location of file to parse}
}
\value{
a tibble
}
\description{
\code{read_psms()} get search results from a variety of platforms
}
