% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_sage.R
\name{read_sage}
\alias{read_sage}
\title{helper function to read in platform specific results}
\usage{
read_sage(x, cpus = 1)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
