% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Table of spectrum values}
\usage{
read_spectra(path = NULL, include_spectra = TRUE)
}
\arguments{
\item{path}{location of file to parse}

\item{include_spectra}{boolean to keep/drop the nested MSn spectrum}
}
\value{
a tibble
}
\description{
\code{read_spectra()} get search results from a variety of platforms
}
