% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_spectrum.R
\name{select_scans}
\alias{select_scans}
\title{helper function to sub select specific results}
\usage{
select_scans(data = NULL, scans = NULL)
}
\arguments{
\item{data}{a ms2spectra data object}

\item{scans}{integer values(s) for the ms_scan events to select for}
}
\value{
a tibble
}
\description{
helper function to sub select specific results
}
