% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spectrum_accuracy.R
\name{sim_spectrum_accuracy}
\alias{sim_spectrum_accuracy}
\title{A helper function used to simulate the accuracy of synthetic spectra.}
\usage{
sim_spectrum_accuracy(df = NULL, accuracy = 0.1, model = stats::rnorm)
}
\arguments{
\item{df}{A dataframe of monoisotopic peaks.}

\item{accuracy}{the targeted mass accuracy defined as the mean of the absolute
mass differences from predicted. Note, that a median mass measurement
accuracy (MMA) of 0.1Th will contain 95\% of all peaks within +/-0.3Th, and
99\% of all peaks within +/-0.4Th, therefor with any given MMA, tolerances
should account for 3x or 4x, to capture 95\% or 99\% of all peaks, respectively.}

\item{model}{A the statistical sampling method to use of assessing the match accuracy}
}
\description{
\code{sim_spectrum_accuracy()} Adds the noise to the accuracy of a synthetic spectrum.
}
