% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_denoise.R
\name{spectrum_denoise}
\alias{spectrum_denoise}
\title{Clean up a mass spectrum to expose only the most abundant peaks}
\usage{
spectrum_denoise(spectrum = NULL, precursor = NULL, hedge = 5, n = 30)
}
\arguments{
\item{spectrum}{a spectrum data object}

\item{precursor}{the precursor mass to remove}

\item{hedge}{the mz space to remove on either side of a tall peak}

\item{n}{the number of peaks to retain}
}
\description{
\code{denoise_spectrum()} Removes the noise from a given mass spectrum
}
