% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_extract.R
\name{spectrum_extract}
\alias{spectrum_extract}
\title{A helper function to extract just the spectrum of the spectra data object}
\usage{
spectrum_extract(spectra = NULL, precursor = FALSE, isotopes = FALSE)
}
\arguments{
\item{spectra}{A spectra data object}

\item{precursor}{A boolean to indicate spectrum filtering should happen}

\item{isotopes}{A boolean to remove isotopes.}
}
\description{
A helper function to extract just the spectrum of the spectra data object
}
