% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_isotopes.R
\name{spectrum_isotopes}
\alias{spectrum_isotopes}
\title{Define the isotopes within a spectrum}
\usage{
spectrum_isotopes(spectrum = NULL)
}
\arguments{
\item{spectrum}{A spectrum data object.}
}
\description{
\code{spectrum_isotopes()} assigns the predicted fragment masses of a given peptide
sequence with the mass spectrum.
}
\examples{
 tbl <- path_to_example() |>
         readr::read_csv()
 tbl |>
   spectrum_isotopes()

}
