% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_noise.R
\name{spectrum_noise}
\alias{spectrum_noise}
\title{Convert a peptide string to a named variable}
\usage{
spectrum_noise(
  df = NULL,
  n_peaks = 300,
  noise_mean = NULL,
  noise_quantile = 0.25,
  noise_sd = 0.75
)
}
\arguments{
\item{df}{a data frame}

\item{n_peaks}{total number of peaks, including noise, the spectrum should have}

\item{noise_mean}{the level at which to inject noise, left NULL the Nth quantile will be used}

\item{noise_quantile}{the Nth quantile used to assess noise}

\item{noise_sd}{the fraction of the noise_quantile to use as the noise variability}
}
\value{
a data frame
}
\description{
\code{spectrum_noise()} Generates a named variable
}
