% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_simulate.R
\name{spectrum_simulate}
\alias{spectrum_simulate}
\title{Convert a peptide string to a named variable}
\usage{
spectrum_simulate(
  peptide = NULL,
  charge = 1:2,
  accuracy = 0.1,
  noise_mean = NULL,
  noise_quantile = 0.33,
  f_peaks = 0.67
)
}
\arguments{
\item{data}{file path}
}
\value{
data object
}
\description{
\code{generate()} Generates a named variable
}
