% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sequence.R
\name{str_sequence}
\alias{str_sequence}
\title{iteratable function to clean up comet sequence strings}
\usage{
str_sequence(sequences = NULL)
}
\arguments{
\item{sequence}{a vector of character string}
}
\value{
a string vector
}
\description{
\code{str_sequence()} get the mass of a poly amino acid
}
