% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mfg.R
\name{write_mgf}
\alias{write_mgf}
\title{The main function for parsing a fasta file}
\usage{
write_mgf(data = NULL, path = NULL)
}
\arguments{
\item{path}{a character string of the path to the MGF formatted file}
}
\value{
a tibble
}
\description{
\code{read_mgf()} get the current regex
}
