% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_xtandem.R
\name{xtandem_peptide}
\alias{xtandem_peptide}
\title{helper function to read in platform specific results}
\usage{
xtandem_peptide(sequence = NULL, modifications = NULL, val_start = 1)
}
\arguments{
\item{sequence}{string}

\item{modifications}{string}
}
\value{
string
}
\description{
helper function to read in platform specific results
}
