% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_precursor}
\alias{filter_precursor}
\title{Return the top n values from a vector}
\usage{
filter_precursor(spectrum = NULL, mz = 30)
}
\arguments{
\item{spectrum}{The spectra data object}

\item{mz}{The m/z value of the precursor to remove}
}
\description{
\code{filter_precursor()} is a helper function to remove residual precursor from
the spectrum
}
