% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_topn}
\alias{filter_topn}
\title{Get the top most abundant peaks from a spectrum}
\usage{
filter_topn(spectrum = NULL, n = 30)
}
\arguments{
\item{spectrum}{The spectra data object}

\item{n}{The n number of top peaks to return}
}
\description{
\code{filter_topn()} is a helper function to get the top n values from a vector
}
