% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_filter.R
\name{filter_xcorr}
\alias{filter_xcorr}
\title{Return a spectrum normalized according to xcorr publication}
\usage{
filter_xcorr(df, chunk = 10, norm = 50)
}
\arguments{
\item{df}{dataframe}

\item{chunk}{number of divisions to normalize across}

\item{norm}{the value to normalize to}
}
\value{
vector
}
\description{
\code{filter_xcorr()} get the top n values from a vector
}
