% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{fragments}
\alias{fragments}
\title{Get a table of fragments}
\usage{
fragments(
  sequence = "SAMPLE",
  type = c("y", "b"),
  charge = c(1, 2, 3, 4),
  loss = NULL
)
}
\arguments{
\item{sequence}{a character string}

\item{type}{a character}

\item{charge}{an integer}

\item{loss}{a character string}
}
\value{
a tibble
}
\description{
\code{fragments()} Generates a table of fragments from a given peptide character string
}
\examples{
library(rpsm)
fragments("SAMPLER")

}
