% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_charged.R
\name{mass_charged}
\alias{mass_charged}
\title{Calculate the m/z of a charged mass}
\usage{
mass_charged(mass, z)
}
\arguments{
\item{mass}{as integer}

\item{z}{an integer}
}
\value{
a vector
}
\description{
\code{charge_mass()} get the mz of a mass with charge
}
