% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/mass_ladder.R
\name{mass_ladder}
\alias{mass_ladder}
\title{Return the mass ladder vector}
\usage{
mass_ladder(sequence = NULL)

mass_ladder(sequence = NULL)
}
\arguments{
\item{sequence}{as character string}
}
\value{
a vector
}
\description{
\code{mass_ladder()} Generates a named variable
}
