% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_ladder_named.R
\name{mass_ladder_named}
\alias{mass_ladder_named}
\title{Convert a peptide string to a named variable}
\usage{
mass_ladder_named(sequence = NULL)
}
\arguments{
\item{sequence}{as character string}
}
\value{
a vector
}
\description{
\code{mass_ladder()} Generates a named variable
}
