% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spectrum_isotopes.R
\name{model_isotopes}
\alias{model_isotopes}
\title{An isotope model using the BRAIN algorithm}
\usage{
model_isotopes(
  averagene = c("N", "A", "R", "D", "C", "E", "Q", "G", "H", "I", "K", "M", "F", "P",
    "S", "T", "W", "Y", "V"),
  max_n = 42
)
}
\arguments{
\item{averagene}{The amino acid to use for a base polypeptide}

\item{max_n}{The maximum number of isotopes to account for}
}
\description{
\code{model_isotopes()} a helper function to generate isotopes
}
