% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_trunc.R
\name{num_trunc}
\alias{num_trunc}
\title{truncate a number to a given decimal value}
\usage{
num_trunc(x, d = 4)
}
\arguments{
\item{x}{the number to truncate}

\item{d}{decimals to truncate}
}
\value{
a numeric vector
}
\description{
\code{num_trunc()} truncate a decimal to a given precision
}
