% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_delta.R
\name{pairwise_delta}
\alias{pairwise_delta}
\title{Get the mass of an amino acid}
\usage{
pairwise_delta(x = NULL, y = NULL, col_values = NULL, col_names = NULL)
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}

\item{col_values}{a character string}

\item{col_names}{a character string}
}
\value{
a tibble
}
\description{
\code{pairwise_delta()} converts an amino acid character to a numeric
}
