% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{GGplot2 object of a ms2 spectrum}
\usage{
plot_spectrum(
  spectrum = NULL,
  peptides = NULL,
  label_size = 3,
  tolerance = 0.1,
  ...
)
}
\arguments{
\item{spectrum}{as data frame cols = \verb{[mz, i]}}

\item{sequence}{as character string}

\item{charge}{as integer}
}
\value{
a string
}
\description{
\code{plot_spectrum()} get the mass of a poly amino acid
}
