% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_denoise.R
\name{spectrum_denoise}
\alias{spectrum_denoise}
\title{A helper function used to clean up a mass spectrum to expose only the most abundant peaks}
\usage{
spectrum_denoise(spectrum = NULL, precursor = NULL, isotopes = FALSE)
}
\arguments{
\item{spectrum}{A spectrum data object.}

\item{precursor}{The floating point value of precursor mass to remove.}

\item{hedge}{The mz space to remove on either side of a tall peak when filtering.}

\item{n}{The number of peaks to retain, ranked by abundance.}
}
\description{
\code{denoise_spectrum()} Removes peaks from a given mass spectrum filtered to remove residual
precursor peaks, isotopes peaks and any low-level noise.
}
