% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_extract.R
\name{spectrum_extract}
\alias{spectrum_extract}
\title{helper function to read in platform specific results}
\usage{
spectrum_extract(x, filter = FALSE)
}
\arguments{
\item{x}{location of file to parse}
}
\value{
a tibble
}
\description{
helper function to read in platform specific results
}
