% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_extract.R
\name{spectrum_extract}
\alias{spectrum_extract}
\title{A helper function to extract just the spectrum of the spectra data object}
\usage{
spectrum_extract(spectra = NULL, filter = FALSE)
}
\arguments{
\item{spectra}{A spectra data object}

\item{filter}{A boolean to indicate spectrum filtering should happen}
}
\description{
A helper function to extract just the spectrum of the spectra data object
}
