% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_isotopes.R
\name{spectrum_isotopes}
\alias{spectrum_isotopes}
\title{Convert a peptide string to a named variable}
\usage{
spectrum_isotopes(df = NULL, isotope_model = NULL, charge = 1)
}
\arguments{
\item{df}{a dataframe of monoisotopic peaks}

\item{isotope_model}{a model of isotopes for a given polypeptide}

\item{charge}{proton charge to assume for the isotope profile}
}
\value{
data object
}
\description{
\code{spectrum_isotopes()} Generates a named variable
}
