% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_isotopes.R
\name{spectrum_isotopes}
\alias{spectrum_isotopes}
\title{Match peptide predicted fragments with a spectrum.}
\usage{
spectrum_isotopes(spectrum = NULL)
}
\arguments{
\item{spectrum}{A spectrum data object.}
}
\description{
\code{spectrum_isotopes()} assigns the predicted fragment masses of a given peptide
sequence with the mass spectrum.
}
\examples{
 # using the supplied spectrum from the msreadr package
 library(msreadr)
 mzml <- path_to_example() |>
         read_spectra()
 mzml |>
   subset(spectrum_num == 1) |>
   spectrum_extract()

}
