% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_isotopes.R
\name{spectrum_isotopes}
\alias{spectrum_isotopes}
\title{Simulate isotope peaks}
\usage{
spectrum_isotopes(spectrum = NULL, isotope_model = NULL, charge = 1)
}
\arguments{
\item{spectrum}{A dataframe of monoisotopic peaks}

\item{isotope_model}{A model of isotopes for a given polypeptide}

\item{charge}{The proton charge to assume for the isotope profile}
}
\description{
\code{spectrum_isotopes()} is a helper function to generate isotope peaks in
synthetic spectra.
}
