% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_peptide.R
\name{str_clean}
\alias{str_clean}
\title{iteratable function to clean up comet sequence strings}
\usage{
str_clean(sequences = NULL)
}
\arguments{
\item{sequence}{as character string}
}
\value{
a string vector
}
\description{
\code{str_peptide()} get the mass of a poly amino acid
}
